+++ src/test/java/br/com/danielwisky/rinhadebackend/templates/resources/TransactionRequestTemplate.java
@@ -1,14 +1,14 @@
 package br.com.danielwisky.rinhadebackend.templates.resources;
 
 import br.com.danielwisky.rinhadebackend.gateways.input.controller.resources.request.TransactionRequest;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class TransactionRequestTemplate {
 
   public static TransactionRequest validCredit() {
     final TransactionRequest transactionRequest = new TransactionRequest();
     transactionRequest.setType('c');
-    transactionRequest.setValue(BigInteger.valueOf(100));
+    transactionRequest.setValue(BigDecimal.valueOf(100));
     transactionRequest.setDescription("credit");
     return transactionRequest;
   }
@@ -16,7 +16,7 @@ public class TransactionRequestTemplate {
   public static TransactionRequest validDebit() {
     final TransactionRequest transactionRequest = new TransactionRequest();
     transactionRequest.setType('d');
-    transactionRequest.setValue(BigInteger.valueOf(10));
+    transactionRequest.setValue(BigDecimal.valueOf(10));
     transactionRequest.setDescription("debit");
     return transactionRequest;
   }
@@ -24,7 +24,7 @@ public class TransactionRequestTemplate {
   public static TransactionRequest validDebitWithHighValue() {
     final TransactionRequest transactionRequest = new TransactionRequest();
     transactionRequest.setType('d');
-    transactionRequest.setValue(BigInteger.valueOf(1001));
+    transactionRequest.setValue(BigDecimal.valueOf(1001));
     transactionRequest.setDescription("debit");
     return transactionRequest;
   }
